Programs given by mentor

+++ dbform.php
@@ -7,7 +7,7 @@
 First Name: <input type="text" name="firstname"><br>
 Last Name : <input type="text" name="lastname"><br>
 Address : 	<input type="text" name="address"><br>
-Image : 	<input type="file" name="image"><br>
+Image : 	<input type="file" name="image[]" multiple><br>
 Bloodgroup : <select name="bloodgroup">
 				<option name="none"> Select Bloodgroup </option> 
 				<option name="A+"> A+ </option> 
@@ -23,6 +23,7 @@ Bloodgroup : <select name="bloodgroup">
 </form>
 
 <?php
+error_reporting(E_ALL);	
 if(isset($_POST['submit']))
 {
 	$servername="localhost";
@@ -35,28 +36,36 @@ if(isset($_POST['submit']))
 	$image=$_POST['image'];
 	$bloodgroup=$_POST['bloodgroup'];
 	$tbname="form_data";
-
-
+	if (!is_dir("images")) 
+			{	
+				 mkdir("/var/www/html/learning/images");
+			 }
+			 echo "<pre>";
+			 print_r($_FILES);
+			die;
+	
 if (($_FILES['image']['name']!=""))
 		{
-
+			print_r($_FILES);
+			die;
+		 $temp=strtotime("now");
 		 $target_dir = "images/";
 		 $file = $_FILES['image']['name'];
 		 $path = pathinfo($file);
 		 $filename = $path['filename'];
 		 $ext = $path['extension'];
 		 $temp_name = $_FILES['image']['tmp_name'];
-		 echo $path_filename_ext = $target_dir.$filename.".".$ext;
-
-		/*if (file_exists($path_filename_ext)) 
+		 $img_name=$filename.$temp.".".$ext;
+		 $path_filename_ext = $target_dir.$img_name;
+		if (file_exists($path_filename_ext)) 
 			{
 			 echo "Sorry, file already exists.";
-			 }*/
-		 	/*else
-			 {*/
+			 }
+		 	else
+			 {
 			 move_uploaded_file($temp_name,$path_filename_ext);
 			 echo "Congratulations! File Uploaded Successfully.";
-			 /*}*/
+			 }
 
 }
 
@@ -81,18 +90,26 @@ mysqli_select_db($dbname, $conn);
 	{
 		echo "Error in creating table .".mysqli_error($conn);
 	}
-        $sql3="INSERT INTO ".$tbname."(FirstName,LastName,Address,Image,Bloodgroup) values('".$first_name."','".$last_name."','".$address."','".$image."','".$bloodgroup."')";
+        $sql3="INSERT INTO ".$tbname."(FirstName,LastName,Address,Image,Bloodgroup) values('".$first_name."','".$last_name."','".$address."','".$img_name."','".$bloodgroup."')";
 
 	if(mysqli_query($conn,$sql3))
-	{
+	{	
 		echo "Inserted successfully.";
-		//header('Location :'.$url);
+			//header('Location :'.$url);
+
+		$sql4="select * from form_data";
+		$result=mysqli_query($conn,$sql4);
+		while($row=mysqli_fetch_assoc($result))
+		{
+			echo "<div><img src='images/".$row['Image']."' height='110px' width='160px'></div>";
+		}
+		
 	}
 	else
 	{
 		echo "Error in inserting.".mysqli_error($conn);
 	}
-mysqli_close($conn);
+		mysqli_close($conn);
 }
